+++ braintree/local_payment_reversed.py
@@ -1,0 +1,9 @@
+from braintree.resource import Resource
+
+class LocalPaymentReversed(Resource):
+    """
+    A class representing Braintree LocalPaymentReversed webhook.
+    """
+    def __init__(self, gateway, attributes):
+        Resource.__init__(self, gateway, attributes)
+